Created module to handle data
A seperate module(storeDeleteData.js) that will
handle data via web storage

The new module creates project objcts, each object
holding a task list which is a list of all task objects
related to that project. The projects are stored in a list
for better accessibility. Web storage is used to store that
data in a JSON file. Functions to do so are created.